% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.data.R
\name{EstimateProliferativeCell}
\alias{EstimateProliferativeCell}
\title{Estimate proliferative cells}
\usage{
EstimateProliferativeCell(DataMat, EstProlif = "MeanPerc", QuaThr = 0.5)
}
\arguments{
\item{DataMat}{expression matrix (rows are cells and cols are genes)}

\item{EstProlif}{estimation algorithm}

\item{QuaThr}{selection parameter}
}
\description{
Estimate proliferative cells
}
