% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsePeacks.R
\name{ExtractStages}
\alias{ExtractStages}
\title{Title}
\usage{
ExtractStages(PeakList, StageVect = list(G1 = "G1", S = "S", G2M = c("G2M",
  "G2/M")), Mode = "Any")
}
\arguments{
\item{Mode}{}
}
\description{
Title
}
