% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsePeacks.R
\name{GetGenesWithPeaks}
\alias{GetGenesWithPeaks}
\title{Title}
\usage{
GetGenesWithPeaks(DataStruct, ProcStruct, Mode = "SmoothOnCircleNodes",
  MinNodes = 2, Thr = 0.9, QuantSel = 0.75, Span = 0.3, MaxPsCV = 1,
  nCores = 1)
}
\arguments{
\item{SinglePeack}{}
}
\description{
Title
}
