% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotOnStages}
\alias{PlotOnStages}
\title{Title}
\usage{
PlotOnStages(Structure, PrinStruct, ExpData, SelThr = NULL,
  SmoothPoints.Growth = 1, SmoothPoints.Shrink = 1,
  ComputeOverlaps = TRUE, nGenes = 10, OrderOnCat = FALSE,
  MinCellPerNode = 3, Title = "")
}
\arguments{
\item{ExpData}{}
}
\description{
Title
}
