% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_detection.R
\name{ProjectAndCompute}
\alias{ProjectAndCompute}
\title{Title}
\usage{
ProjectAndCompute(DataMat, nNodes = 30, GraphType = "Circle",
  PlanVarLimit = 0.9, PlanVarLimitIC = NULL, InitStructNodes = 15,
  nCores = 1, NonG0Cell)
}
\arguments{
\item{DataMat}{}

\item{nCores}{}

\item{NonG0Cell}{}
}
\description{
Title
}
\examples{

}
