% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneselectiononcurve.R
\name{SelectGenes}
\alias{SelectGenes}
\title{Title}
\usage{
SelectGenes(Partition, ExpMat, Mode = "CV", AggFun = min, Span = 0.75,
  Net = NULL, Edges = NULL, ProjStruct = NULL, nCores = 1)
}
\arguments{
\item{Proj}{}
}
\description{
Title
}
