% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsePeacks.R
\name{StageWithPeaks}
\alias{StageWithPeaks}
\title{Title}
\usage{
StageWithPeaks(DataStruct, ProcStruct, StageInfo.UP, StageInfo.DOWN = list(),
  ComputeG0 = TRUE, CCGenes = NULL, MinNodes = 2, QuantSel = 0.75,
  G0Level = 0.75, CCLevel = 0.95, Title = "",
  StagingMode = "SmoothOnCircleNodes", Mode = 1, Span = 0.2, nCores = 1,
  MaxPsCV = 2, FullStaging = FALSE)
}
\arguments{
\item{Mode}{}
}
\description{
Title
}
